python/agents/1-custom-fun/requirements.txt

google-adk
google-api-python-client
google-auth-oauthlib
google-auth
dateparser
python-dateutil
pytz
tzlocal
